# ArenaREST/Models/Stall.cs
public class Stall
{
    public int StallId { get; set; }
    public string Name { get; set; }
    public string Type { get; set; }
    public int Floor { get; set; }
    public string Coordinates { get; set; } 
}